wrap up LPC launch logic in hook
